+++ examples/esp32/main/my_platform.c
@@ -98,8 +98,9 @@ static void my_platform_on_controller_data(uni_hid_device_t* d, uni_controller_t
 
             // Debugging
             // Axis ry: control rumble
-            if ((gp->buttons & BUTTON_A) && d->report_parser.set_rumble != NULL) {
-                d->report_parser.set_rumble(d, 128, 128);
+            if ((gp->buttons & BUTTON_A) && d->report_parser.play_dual_rumble != NULL) {
+                d->report_parser.play_dual_rumble(d, 0 /* delayed start ms */, 50 /* duration ms */,
+                                                  128 /* weak magnitude */, 40 /* strong magnitude */);
             }
             // Buttons: Control LEDs On/Off
             if ((gp->buttons & BUTTON_B) && d->report_parser.set_player_leds != NULL) {
@@ -173,8 +174,9 @@ static my_platform_instance_t* get_my_platform_instance(uni_hid_device_t* d) {
 static void trigger_event_on_gamepad(uni_hid_device_t* d) {
     my_platform_instance_t* ins = get_my_platform_instance(d);
 
-    if (d->report_parser.set_rumble != NULL) {
-        d->report_parser.set_rumble(d, 0x80 /* value */, 15 /* duration */);
+    if (d->report_parser.play_dual_rumble!= NULL) {
+        d->report_parser.play_dual_rumble(d, 0 /* delayed start ms */, 150 /* duration ms */,
+                                          128 /* weak magnitude */, 40 /* strong magnitude */);
     }
 
     if (d->report_parser.set_player_leds != NULL) {