+++ examples/pico_w/src/my_platform.c
@@ -130,8 +130,9 @@ static void my_platform_on_controller_data(uni_hid_device_t* d, uni_controller_t
 
             // Debugging
             // Axis ry: control rumble
-            if ((gp->buttons & BUTTON_A) && d->report_parser.set_rumble != NULL) {
-                d->report_parser.set_rumble(d, 128, 128);
+            if ((gp->buttons & BUTTON_A) && d->report_parser.play_dual_rumble != NULL) {
+                d->report_parser.play_dual_rumble(d, 0 /* delayed start ms */, 150 /* duration ms */,
+                                                  128 /* weak magnitude */, 40 /* strong magnitude */);
             }
             // Buttons: Control LEDs On/Off
             if ((gp->buttons & BUTTON_B) && d->report_parser.set_player_leds != NULL) {
@@ -233,8 +234,9 @@ static void my_platform_on_oob_event(uni_platform_oob_event_t event, void* data)
 // Helpers
 //
 static void trigger_event_on_gamepad(uni_hid_device_t* d) {
-    if (d->report_parser.set_rumble != NULL) {
-        d->report_parser.set_rumble(d, 0x80 /* value */, 15 /* duration */);
+    if (d->report_parser.play_dual_rumble != NULL) {
+        d->report_parser.play_dual_rumble(d, 0 /* delayed start ms */, 50 /* duration ms */, 128 /* weak magnitude */,
+                                          40 /* strong magnitude */);
     }
 
     if (d->report_parser.set_player_leds != NULL) {