+++ examples/posix/src/my_platform.c
@@ -112,8 +112,6 @@ static void posix_on_controller_data(uni_hid_device_t* d, uni_controller_t* ctl)
             // Axis ry: control rumble
             if ((gp->buttons & BUTTON_A) && d->report_parser.play_dual_rumble != NULL) {
                 d->report_parser.play_dual_rumble(d, 1000, 100, 0, 255);
-            } else if ((gp->buttons & BUTTON_A) && d->report_parser.set_rumble != NULL) {
-                d->report_parser.set_rumble(d, 0x80, 50);
             }
 
             if ((gp->buttons & BUTTON_B) && d->report_parser.play_dual_rumble != NULL) {
@@ -208,8 +206,6 @@ static void trigger_event_on_gamepad(uni_hid_device_t* d) {
         d->report_parser.play_dual_rumble(
             d, 0 /* delayed start ms */, 150 /* duration ms */, 0 /* weak_magnitude */, 255 /* strong_magnitude */
         );
-    } else if (d->report_parser.set_rumble != NULL) {
-        d->report_parser.set_rumble(d, 0x80 /* value */, 15 /* duration */);
     }
 
     if (d->report_parser.set_player_leds != NULL) {