+++ src/components/bluepad32/include/parser/uni_hid_parser_ds3.h
@@ -15,7 +15,6 @@ void uni_hid_parser_ds3_setup(struct uni_hid_device_s* d);
 void uni_hid_parser_ds3_init_report(struct uni_hid_device_s* d);
 void uni_hid_parser_ds3_parse_input_report(struct uni_hid_device_s* d, const uint8_t* report, uint16_t len);
 void uni_hid_parser_ds3_set_player_leds(struct uni_hid_device_s* d, uint8_t leds);
-void uni_hid_parser_ds3_set_rumble(struct uni_hid_device_s* d, uint8_t value, uint8_t duration);
 void uni_hid_parser_ds3_play_dual_rumble(struct uni_hid_device_s* d,
                                          uint16_t start_delay_ms,
                                          uint16_t duration_ms,