+++ src/components/bluepad32/include/parser/uni_hid_parser_ds4.h
@@ -15,7 +15,6 @@ void uni_hid_parser_ds4_init_report(struct uni_hid_device_s* d);
 void uni_hid_parser_ds4_parse_input_report(struct uni_hid_device_s* d, const uint8_t* report, uint16_t len);
 void uni_hid_parser_ds4_parse_feature_report(struct uni_hid_device_s* d, const uint8_t* report, uint16_t len);
 void uni_hid_parser_ds4_set_lightbar_color(struct uni_hid_device_s* d, uint8_t r, uint8_t g, uint8_t b);
-void uni_hid_parser_ds4_set_rumble(struct uni_hid_device_s* d, uint8_t value, uint8_t duration);
 void uni_hid_parser_ds4_play_dual_rumble(struct uni_hid_device_s* d,
                                          uint16_t start_delay_ms,
                                          uint16_t duration_ms,