+++ src/components/bluepad32/include/parser/uni_hid_parser_ds5.h
@@ -56,7 +56,6 @@ void uni_hid_parser_ds5_parse_input_report(struct uni_hid_device_s* d, const uin
 void uni_hid_parser_ds5_parse_feature_report(struct uni_hid_device_s* d, const uint8_t* report, uint16_t len);
 void uni_hid_parser_ds5_set_player_leds(struct uni_hid_device_s* d, uint8_t value);
 void uni_hid_parser_ds5_set_lightbar_color(struct uni_hid_device_s* d, uint8_t r, uint8_t g, uint8_t b);
-void uni_hid_parser_ds5_set_rumble(struct uni_hid_device_s* d, uint8_t value, uint8_t duration);
 void uni_hid_parser_ds5_play_dual_rumble(struct uni_hid_device_s* d,
                                          uint16_t start_delay_ms,
                                          uint16_t duration_ms,