+++ src/components/bluepad32/include/parser/uni_hid_parser_psmove.h
@@ -14,7 +14,11 @@
 void uni_hid_parser_psmove_setup(struct uni_hid_device_s* d);
 void uni_hid_parser_psmove_init_report(struct uni_hid_device_s* d);
 void uni_hid_parser_psmove_parse_input_report(struct uni_hid_device_s* d, const uint8_t* report, uint16_t len);
-void uni_hid_parser_psmove_set_rumble(struct uni_hid_device_s* d, uint8_t value, uint8_t duration);
 void uni_hid_parser_psmove_set_lightbar_color(struct uni_hid_device_s* d, uint8_t r, uint8_t g, uint8_t b);
+void uni_hid_parser_psmove_play_dual_rumble(struct uni_hid_device_s* d,
+                                            uint16_t start_delay_ms,
+                                            uint16_t duration_ms,
+                                            uint8_t weak_magnitude,
+                                            uint8_t strong_magnitude);
 
 #endif  // UNI_HID_PARSER_PSMOVE_H