+++ src/components/bluepad32/include/parser/uni_hid_parser_stadia.h
@@ -13,7 +13,6 @@
 #define UNI_HID_PARSER_STADIA_PID 0x9400
 
 void uni_hid_parser_stadia_setup(struct uni_hid_device_s* d);
-void uni_hid_parser_stadia_set_rumble(struct uni_hid_device_s* d, uint8_t value, uint8_t duration);
 void uni_hid_parser_stadia_play_dual_rumble(struct uni_hid_device_s* d,
                                             uint16_t start_delay_ms,
                                             uint16_t duration_ms,