+++ src/components/bluepad32/parser/uni_hid_parser_ds5.c
@@ -678,10 +678,6 @@ void uni_hid_parser_ds5_set_lightbar_color(struct uni_hid_device_s* d, uint8_t r
     ds5_send_output_report(d, &out);
 }
 
-void uni_hid_parser_ds5_set_rumble(struct uni_hid_device_s* d, uint8_t value, uint8_t duration) {
-    uni_hid_parser_ds5_play_dual_rumble(d, 0, duration * 4, value, value);
-}
-
 void uni_hid_parser_ds5_play_dual_rumble(struct uni_hid_device_s* d,
                                          uint16_t start_delay_ms,
                                          uint16_t duration_ms,