+++ src/components/bluepad32/parser/uni_hid_parser_switch.c
@@ -1116,10 +1116,6 @@ void uni_hid_parser_switch_set_player_leds(uni_hid_device_t* d, uint8_t leds) {
     set_led(d, leds);
 }
 
-void uni_hid_parser_switch_set_rumble(struct uni_hid_device_s* d, uint8_t value, uint8_t duration) {
-    uni_hid_parser_switch_play_dual_rumble(d, 0, duration * 4, value, value);
-}
-
 void uni_hid_parser_switch_play_dual_rumble(struct uni_hid_device_s* d,
                                             uint16_t start_delay_ms,
                                             uint16_t duration_ms,